test: fix some `i18n` tests and move tests from `ui`
